MAINT: Move DofManager to dedicated file
